+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [](https://github.com/juanbrujo/CrawlMapper/compare/v2.3.0...v) (2025-12-15)
+
+### Bug Fixes
+
+- add startup message to server for better UX ([3511060](https://github.com/juanbrujo/CrawlMapper/commit/35110602f93be57229356811f7efe84e15a5729e))
+
 ### [2.2.1](https://github.com/juanbrujo/CrawlMapper/compare/v2.2.0...v2.2.1) (2025-12-15)
 
 ### Features